removed intent.py and moved function to telegram controller
